+++ src/app/(logged-in)/my/legal/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {notFound} from 'next/navigation';
 import {getSession} from '@auth0/nextjs-auth0';
-import {getAllCorporationTypes} from '@/lib/get-all-corporation-types.ts';
+import {getAllCorporationTypes} from '@/lib/models/corporationType.ts';
 import LegalInfoForm from '@/app/(logged-in)/my/legal/legal-info-form.tsx';
 import updateOrganizationAction from '@/lib/actions/update-organization-action.ts';
 import prisma from '@/lib/prisma.ts';
