+++ src/app/api/sectors/route.ts
@@ -1,5 +1,5 @@
 import {type NextRequest, NextResponse} from 'next/server';
-import getAllSectors from '@/lib/get-all-sectors.ts';
+import {getAllSectors} from '@/lib/models/sector';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export const GET = async (request: NextRequest) => {
