--- src/lib/get-all-age-groups.ts
@@ -1,4 +1,0 @@
-import {cache} from 'react';
-import prisma from '@/lib/prisma.ts';
-
-export const getAllAgeGroups = cache(async () => prisma.ageGroup.findMany());
