+++ src/lib/models/ageGroup.ts
@@ -1,0 +1,4 @@
+import {cache} from 'react';
+import prisma from '@/lib/prisma.ts';
+
+export const getAllAgeGroups = cache(async () => prisma.ageGroup.findMany());
