+++ src/lib/models/employeeCountCategory.ts
@@ -1,7 +1,7 @@
 import {cache} from 'react';
 import prisma from '@/lib/prisma.ts';
 
-export default cache(async () => prisma.volunteerCountCategory.findMany({
+export const getAllEmployeeCountCategories = cache(async () => prisma.employeeCountCategory.findMany({
 	orderBy: {
 		minCount: 'asc',
 	},
