+++ src/lib/models/sector.ts
@@ -3,7 +3,7 @@ import {type Sector} from '@prisma/client';
 import {type Geometry} from 'geojson';
 import prisma from '@/lib/prisma.ts';
 
-export default cache(async () => {
+export const getAllSectors = cache(async () => {
 	const sectors: Array<Sector & {geom: Geometry; municipalityName: string}> = await prisma.$queryRaw`select s.id,
                                                                                   s.name,
                                                                                   s."municipalityId",
